platform: Factor out common stm32 buffer handling/swap code
This will be shared by gd32 as well
